・プロフィールモーダルに表示する対象ユーザーの好き登録動画について、最初に10件分表示し、最下部までスクロールすると新たに10件取得して連…

+++ src/Controller/MoviesController.php
@@ -7,35 +7,49 @@ class MoviesController extends AppController {
   public function initialize() {
     $this->name = 'Movies';
     $this->autoRender = false;
+    session_start();
   }
 
   public function store() {
-    $movie = $this->Movies->newEntity();
-    $movie->link = $this->request->getData('link');
-    $movie->title = $this->request->getData('title');
-    $movie->description = $this->request->getData('description');
-    $movie->channel_title = $this->request->getData('channelTitle');
-    $movie->view_count = $this->request->getData('viewCount');
-    $movie->like_count = $this->request->getData('likeCount');
-    $movie->dislike_count = $this->request->getData('dislikeCount');
-    $movie->comment_count = $this->request->getData('commentCount');
-    $movie->group_name = $this->request->getData('groupName');
+    // ログインチェック
+    if (isset($_SESSION['name'])) {
+      if (!($_SESSION['id'] == 1079030440323833857)) {
+        $this->redirect(['controller'=>'top','action','index']);
+      } else {
+        $movie = $this->Movies->newEntity();
+        $movie->link = $this->request->getData('link');
+        $movie->title = $this->request->getData('title');
+        $movie->description = $this->request->getData('description');
+        $movie->channel_title = $this->request->getData('channelTitle');
+        $movie->view_count = $this->request->getData('viewCount');
+        $movie->like_count = $this->request->getData('likeCount');
+        $movie->dislike_count = $this->request->getData('dislikeCount');
+        $movie->comment_count = $this->request->getData('commentCount');
+        $movie->group_name = $this->request->getData('groupName');
 
-    if ($this->Movies->save($movie)) {
-      $this->log('ok');
+        if ($this->Movies->save($movie)) {
+          $this->log('ok');
+        }
+      }
+    } else {
+      $this->redirect(['controller'=>'top','action'=>'index']);
     }
   }
 
   public function delete() {
-    $id = $this->request->getData('movie-delete-hidden');
-    $movie = $this->Movies->get($id);
-    $this->Movies->delete($movie);
-    $this->redirect(['controller'=>'Top','action'=>'index']);
+    if (isset($_SESSION['name'])) {
+      if (!($_SESSION['id'] == 1079030440323833857)) {
+        $this->redirect(['controller'=>'top','action','index']);
+      } else {
+        $id = $this->request->getData('movie-delete-hidden');
+        $this->log($id);
+        $movie = $this->Movies->get($id);
+        $this->Movies->delete($movie);
+        $this->redirect(['controller'=>'Top','action'=>'index']);
+      }
+    } else {
+      $this->redirect(['controller'=>'top','action'=>'index']);
+    }
   }
 
-  public function likestore() {
-    $this->autoRender = false;
-    $entity = $this->Movies->find()->contain(['Users'])->all();
-    dd($entity);
-  }
 }
